remove old project and rename
