+++ SWKOM-paperless.sln
@@ -1,8 +1,6 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SWKOM-paperless", "SWKOM-paperless\SWKOM-paperless.csproj", "{009E85CC-D88A-4A00-88DA-4BB14BE14772}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Org.OpenAPITools", "Org.OpenAPITools\Org.OpenAPITools.csproj", "{96F2E125-CD32-4FB3-AB72-4E1F7F35159F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SWKOM-paperless", "SWKOM-paperless\SWKOM-paperless.csproj", "{96F2E125-CD32-4FB3-AB72-4E1F7F35159F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -10,10 +8,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{009E85CC-D88A-4A00-88DA-4BB14BE14772}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{009E85CC-D88A-4A00-88DA-4BB14BE14772}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{009E85CC-D88A-4A00-88DA-4BB14BE14772}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{009E85CC-D88A-4A00-88DA-4BB14BE14772}.Release|Any CPU.Build.0 = Release|Any CPU
 		{96F2E125-CD32-4FB3-AB72-4E1F7F35159F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{96F2E125-CD32-4FB3-AB72-4E1F7F35159F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{96F2E125-CD32-4FB3-AB72-4E1F7F35159F}.Release|Any CPU.ActiveCfg = Release|Any CPU
