+++ SWKOM-paperless/.gitignore
