+++ SWKOM-paperless/Authentication/ApiAuthentication.cs
