+++ SWKOM-paperless/Controllers/SearchApi.cs
