+++ SWKOM-paperless/Converters/CustomEnumConverter.cs
