+++ SWKOM-paperless/Filters/BasePathFilter.cs
