+++ SWKOM-paperless/Formatters/InputFormatterStream.cs
