+++ SWKOM-paperless/Models/Document.cs
