+++ SWKOM-paperless/Models/NewDocumentType.cs
