+++ SWKOM-paperless/Models/UserInfo.cs
