+++ SWKOM-paperless/OpenApi/TypeExtensions.cs
