+++ SWKOM-paperless/wwwroot/README.md
